feat(dark-theme): apply dark theme to workout page
re #53
